Split components into dedicated files
