+++ Sources/HTMLKitComponents/Components/HStack.swift
@@ -1,0 +1,120 @@
+import HTMLKit
+
+/// A component that arranges content horizontally.
+public struct HStack: View, Actionable, Modifiable {
+    
+    /// The identifier of the text.
+    internal var id: String?
+    
+    /// The content of the stack.
+    internal var content: [Content]
+    
+    /// The classes of the stack.
+    internal var classes: [String]
+    
+    /// The events of the text.
+    internal var events: [String]?
+     
+    /// Creates a horizontal stack.
+    public init(alignment: Tokens.VerticalAlignment = .center, spacing: Tokens.ContentSpace? = nil, @ContentBuilder<Content> content: () -> [Content]) {
+        
+        self.content = content()
+        
+        if let spacing {
+            self.classes = ["hstack", "vertical-alignment:\(alignment.value)", "spacing:\(spacing.value)"]
+            
+        } else {
+            self.classes = ["hstack", "vertical-alignment:\(alignment.value)"]
+        }
+    }
+    
+    public var body: Content {
+        Division {
+            content
+        }
+        .class(classes.joined(separator: " "))
+        .modify(unwrap: id) {
+            $0.id($1)
+        }
+        if let events = self.events {
+            Script {
+                events
+            }
+        }
+    }
+    
+    public func tag(_ value: String) -> HStack {
+        return self.mutate(id: value)
+    }
+    
+    public func shadow(_ radius: Tokens.BlurRadius = .small, color: Tokens.ShadowColor = .black) -> HStack {
+        return self.mutate(classes: ["shadow:\(radius.value)", "shadow:\(color.value)"])
+    }
+    
+    public func clipped() -> HStack {
+        return self.mutate(class: "overflow:clip")
+    }
+}
+
+extension HStack: MouseEvent {
+    
+    public func onHover(@ActionBuilder action: (ViewAction) -> [Action]) -> HStack {
+        return self.mutate(hoverevent: action(.init()))
+    }
+    
+    public func onLeave(@ActionBuilder action: (ViewAction) -> [Action]) -> HStack {
+        return self.mutate(leaveevent: action(.init()))
+    }
+}
+
+extension HStack: ViewModifier {
+
+    public func backgroundColor(_ color: Tokens.BackgroundColor) -> HStack {
+        return self.mutate(backgroundcolor: color.value)
+    }
+
+    public func opacity(_ value: Tokens.OpacityValue) -> HStack {
+        return self.mutate(opacity: value.value)
+    }
+    
+    public func zIndex(_ index: Tokens.PositionIndex) -> HStack {
+        return self.mutate(zindex: index.value)
+    }
+    
+    public func hidden() -> HStack {
+        return self.mutate(viewstate: Tokens.ViewState.hidden.value)
+    }
+    
+    public func hidden(_ condition: Bool) -> HStack {
+        
+        if condition {
+            return self.mutate(viewstate: Tokens.ViewState.hidden.value)
+        }
+        
+        return self
+    }
+    
+    public func colorScheme(_ scheme: Tokens.ColorScheme) -> HStack {
+        return self.mutate(scheme: scheme.value)
+    }
+    
+    public func padding(insets: EdgeSet = .all, length: Tokens.PaddingLength = .small) -> HStack {
+        return self.mutate(padding: length.value, insets: insets)
+    }
+    
+    public func borderShape(_ shape: Tokens.BorderShape) -> HStack {
+        return self.mutate(bordershape: shape.value)
+    }
+    
+    public func border(_ color: Tokens.BorderColor, width: Tokens.BorderWidth = .small) -> HStack {
+        return self.mutate(border: color.value, width: width.value)
+    }
+    
+    public func frame(width: Tokens.ViewWidth, height: Tokens.ViewHeight? = nil, alignment: Tokens.FrameAlignment? = nil) -> HStack {
+        return mutate(frame: width.value, height: height?.value, alignment: alignment?.value)
+    }
+    
+    public func margin(insets: EdgeSet = .all, length: Tokens.MarginLength = .small) -> HStack {
+        return self.mutate(margin: length.value, insets: insets)
+    }
+}
