+++ Sources/HTMLKitComponents/Components/Slider.swift
@@ -1,0 +1,105 @@
+import HTMLKit
+
+/// A component that displays
+public struct Slider: View, Modifiable, Identifiable {
+    
+    internal var id: String?
+    
+    /// The identifier of the slider.
+    internal let name: String
+    
+    /// The classes of the slider.
+    internal var classes: [String]
+    
+    /// The events of the slider.
+    internal var events: [String]?
+    
+    /// Creates a slider.
+    public init(name: String) {
+        
+        self.name = name
+        self.classes = ["slider"]
+    }
+    
+    public var body: Content {
+        Input()
+            .type(.range)
+            .name(name)
+            .class(classes.joined(separator: " "))
+            .modify(unwrap: id) {
+                $0.id($1)
+            }
+    }
+    
+    public func tag(_ value: String) -> Slider {
+        return self.mutate(id: value)
+    }
+}
+
+extension Slider: InputModifier {
+    
+    public func disabled(_ condition: Bool) -> Slider {
+        
+        if condition {
+            return self.mutate(inputstate: Tokens.ViewState.disabled.value)
+        }
+        
+        return self
+    }
+    
+    public func focusColor(_ color: Tokens.FocusColor) -> Slider {
+        return self.mutate(focuscolor: color.value)
+    }
+}
+
+extension Slider: ViewModifier {
+    
+    public func opacity(_ value: Tokens.OpacityValue) -> Slider {
+        return self.mutate(opacity: value.value)
+    }
+    
+    public func zIndex(_ index: Tokens.PositionIndex) -> Slider {
+        return self.mutate(zindex: index.value)
+    }
+    
+    public func hidden() -> Slider {
+        return self.mutate(viewstate: Tokens.ViewState.hidden.value)
+    }
+    
+    public func hidden(_ condition: Bool) -> Slider {
+        
+        if condition {
+            return self.mutate(viewstate: Tokens.ViewState.hidden.value)
+        }
+        
+        return self
+    }
+    
+    public func padding(insets: EdgeSet = .all, length: Tokens.PaddingLength = .small) -> Slider {
+        return self.mutate(padding: length.value, insets: insets)
+    }
+    
+    public func border(_ color: Tokens.BorderColor, width: Tokens.BorderWidth = .small) -> Slider {
+        return self.mutate(border: color.value, width: width.value)
+    }
+    
+    public func borderShape(_ shape: Tokens.BorderShape) -> Slider {
+        return self.mutate(bordershape: shape.value)
+    }
+    
+    public func backgroundColor(_ color: Tokens.BackgroundColor) -> Slider {
+        return self.mutate(backgroundcolor: color.value)
+    }
+    
+    public func colorScheme(_ scheme: Tokens.ColorScheme) -> Slider {
+        return self.mutate(scheme: scheme.value)
+    }
+    
+    public func frame(width: Tokens.ViewWidth, height: Tokens.ViewHeight? = nil, alignment: Tokens.FrameAlignment? = nil) -> Slider {
+        return mutate(frame: width.value, height: height?.value, alignment: alignment?.value)
+    }
+    
+    public func margin(insets: EdgeSet = .all, length: Tokens.MarginLength = .small) -> Slider {
+        return self.mutate(margin: length.value, insets: insets)
+    }
+}
